requirements.txt
appdirs==1.4.2
defusedxml==0.5.0
dj-database-url==0.4.2
Django==1.10.6
django-registration==2.2
django-smart-selects==1.3.4
django-xor-formfields==0.1.0
gunicorn==19.7.1
oauthlib==2.0.2
packaging==16.8
PyJWT==1.4.2
pyparsing==2.1.10
python-social-auth==0.2.1
python3-openid==3.1.0
requests==2.13.0
requests-oauthlib==0.3.1
six==1.10.0
whitenoise==3.3.0
